refactor: move lookups impl to service
